+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Can now call `abort` on files.createWriteStream to delete partial file
+
 ## [3.0.0-alpha.10] - 2016-02-022
 ### Fixed
 * Rebuild